common/requirements.txt
## gooey-gpu
Pillow ~= 9.4.0
fastapi ~= 0.89.1
uvicorn[standard] ~= 0.20.0
sentry-sdk ~= 1.14.0
scipy ~= 1.10.0
numpy ~= 1.24.3
accelerate ~= 0.17.0

## pytorch
--extra-index-url https://download.pytorch.org/whl/cu118
torch ~= 2.0.0
torchvision ~= 0.15.1
torchaudio ~= 2.0.1

## huggingface diffusers
diffusers ~= 0.16.0
transformers ~= 4.28.1
sentencepiece ~= 0.1.97

## controlnet
controlnet_aux ~= 0.0.1

## bark
git+https://github.com/suno-ai/bark.git
